Support 0.25 increment drag stops
Agent-Logs-Url: https://github.com/enijburg/ManagementMesh/sessions/3eafb949-ef84-47bf-b540-16289f72eac5

+++ src/ManagementMesh.Editor/Models/MeshModels.cs
@@ -21,13 +21,13 @@ public sealed class ManagementMeshDocument
 
     // Vertical threads: independent bars from the top (Start) and bottom (End)
     public Dictionary<string, string> VerticalStartColors { get; set; } = new();
-    public Dictionary<string, int>    VerticalStartStops  { get; set; } = new();
+    public Dictionary<string, double> VerticalStartStops  { get; set; } = new();
     public Dictionary<string, string> VerticalEndColors   { get; set; } = new();
-    public Dictionary<string, int>    VerticalEndStops    { get; set; } = new();
+    public Dictionary<string, double> VerticalEndStops    { get; set; } = new();
 
     // Horizontal threads: independent bars from the left (Start) and right (End)
     public Dictionary<string, string> HorizontalStartColors { get; set; } = new();
-    public Dictionary<string, int>    HorizontalStartStops  { get; set; } = new();
+    public Dictionary<string, double> HorizontalStartStops  { get; set; } = new();
     public Dictionary<string, string> HorizontalEndColors   { get; set; } = new();
-    public Dictionary<string, int>    HorizontalEndStops    { get; set; } = new();
+    public Dictionary<string, double> HorizontalEndStops    { get; set; } = new();
 }
